Dynamic Value for useNativeDriver
Added the ability to have a dynamic value for the useNativeDriver on the Animation.timing of the speedometer image. This will remove the warning from a react native or expo project, and still allow for flexibility of users who may wish to disable this for specific devices. 

defaultProp of useNativeDriver prop has been set to true.

PropType of useNativeDriver prop has been set to boolean.

this commit aims to solve the issue in the creator's repository.

https://github.com/pritishvaidya/react-native-speedometer/issues/29